chore: Add structure to automatically reset between test runs
